requirements.txt
# Gemini API 客户端
google-genai>=1.0.0

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 模板引擎
jinja2>=3.1.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.0.0
